Change Freq to mimic Unibotics
