+++ react_frontend/src/components/visualizers/Frequencies.js
@@ -1,11 +1,19 @@
 import { Typography } from "@mui/material";
 import React, { useState } from "react";
 import "../../styles/visualizers/Frequencies.css";
+import monitor from "../../images/monitoring2.png";
 
-const Frequencies = (props) => {
+const Frequencies = () => {
   const [frequencies, setFrequencies] = useState({ brain: 0, gui: 0, rtf: 0 });
   const [rosVersion, setRosVersion] = useState(null);
   const [gpuAvaliable, setGpuAvaliable] = useState(false);
+  const [showFrequencies, setShowFrequencies] = React.useState(false);
+  const [buttonActive, setButtonActive] = React.useState(false);
+
+  const handleToggleFrequencies = () => {
+    setButtonActive(!buttonActive);
+    setShowFrequencies(!showFrequencies);
+  };
 
   React.useEffect(() => {
     const callback = (message) => {
@@ -44,16 +52,25 @@ const Frequencies = (props) => {
   }, []);
 
   return (
-    <div className={props.style}>
-      <Typography title="BRAIN">{frequencies.brain.toFixed(0)}</Typography>
-      <Typography>/</Typography>
-      <Typography title="RTF">{frequencies.rtf}</Typography>
-      <Typography>/</Typography>
-      <Typography>{rosVersion}</Typography>
-      <Typography>/</Typography>
-      <Typography>GPU</Typography>
-      <Typography>{gpuAvaliable ? "ON" : "OFF"}</Typography>
-    </div>
+    <>
+      <div className={showFrequencies ? "visible" : "hidden"}>
+        <Typography title="BRAIN">{frequencies.brain.toFixed(0)}</Typography>
+        <Typography>/</Typography>
+        <Typography title="RTF">{frequencies.rtf}</Typography>
+        <Typography>/</Typography>
+        <Typography>{rosVersion}</Typography>
+        <Typography>/</Typography>
+        <Typography>GPU</Typography>
+        <Typography>{gpuAvaliable ? "ON" : "OFF"}</Typography>
+      </div>
+      <button
+        className={`button ${buttonActive ? "toggledColor" : ""}`}
+        onClick={handleToggleFrequencies}
+        id="toggleButton"
+      >
+        <img src={monitor} className="monitor"></img>
+      </button>
+    </>
   );
 };
